monitoring/duty_exporter/requirements.txt
prometheus-client==0.17.1
requests==2.31.0
mysql-connector-python==8.2.0